feat: Add `addCQIDsColumns` util, split MemDB into files
