+++ .eslintrc
@@ -22,7 +22,6 @@
     "unicorn/no-null": 0,
     "unused-imports/no-unused-imports": "error",
     "no-console": "error",
-    "@typescript-eslint/no-unused-vars": 0,
     "require-await": "off",
     "@typescript-eslint/require-await": "error",
     "@typescript-eslint/naming-convention": "error",
