--- src/arrow/index.test.ts
@@ -1,8 +1,0 @@
-import test from 'ava';
-
-import { testArrow } from './index.js';
-
-test('testArrow', (t) => {
-  const vectors = testArrow();
-  t.not(vectors, undefined);
-});
